Trim down vim config
Vim has turned into more of a system file editor for me, as opposed
to being my daily driver editor. As such, it's more important for
me that it starts fast, than that it is full featured. This removes
a bunch of plugins that were unused, or just wouldn't matter much
if I'm not using vim all the time.