Emacs bankruptcy
This has honestly been coming on for about a month now.
Spacemacs is just not what it used to be, at least so far as
I'm concerned. When I started using it, Spacemacs was like a
starter pack for Vim users interested in Emacs, it was
opinionated, but I still felt like I had control of it when it
came down to it. Increasingly though, it's feeling like the
Spacemacs team wants to turn it into it's own experience inside
of Emacs, rather than being a starter pack.

I'm frequently running into weird error messages and behaviors
and I have no idea why it's doing these things. I don't think
that Spacemacs is bad, but it's just not for me anymore. And
if I'm being honest with myself, I didn't take my own advice
of not using other people's configs. I may very well come back
and learn Emacs on my own, not relying on someone other
configuration. But for now, there's no reason to keep this around
because it's completely dependent on Spacemacs. I've backed it up
in a gist on GitHub, so that I can pick bits and pieces out when
I come back to it.