Clean up Emacs configuration
Spacemacs does a great job of setting sane defaults for a lot of popular (and
less popular) packages. But the single file configuration was getting to me.
Here I've done something a bit more like my Vim configuration. Where different
parts of my editor configuration live in different files. I think this will make
it a lot easier to keep track of where everything is in the long run.