Fix compilation error with XMonad
This must have had something to do with me having written this on a
different OS (Ubuntu) than the one I just tried to run it on (Fedora)
which, still doesn't make a lot of sense but it's the only difference I
can find.

Spitballing here, but maybe Ubuntu allows XMonad to imply more stuff??
Seriously confusing stuff.