Rework Vim utility plugins
This is mostly jsut alphabetizing, but there is one important change
here. The plugin for TypeScript seems to be unstable, so I've locked it
in to the last version that I was able to use it as expected at. I'll
have to circle back around to it at some point in the future and see if
a newser version has stablized. I added this a while ago, so I know that
commit isn't very fresh.