Use cleaner code to load my own files
Rather than making a bunch of variables that I will never touch after
init, this loads everything in a loop by providing the names of the
files I want out of the directories.

In the future I want to get smarter about this, and only load things as
needed but this is good for now.