Add more bindings bindings for the package menu
The package menu uses the normal Emacs bindings, "n" for next-line,
"p" for previous line, etc. But without the need to hold down ctrl.
However, it's missing a few of the standard bindings, and I find myself
trying to hit things like "l" to recenter often enough that it makes
sense to add it, and some other things I might try to hit from time to
time.