Add functions for not doing things on an OS
I don't actually have a known use for this yet, but I've also yet to
try and run all of this on a Windows machine; and I'm sure that's
going to generate plenty of issues. For example, I'm sure that I'm
going to have to wrap several use-package statements in a
unless-windows-do  function because they're unavailable.