Add a whole bunch of pretty stuff to Emacs
The mode-line, neotree, and dired all use the all-the-icons package
now.

It's worth pointing out that I may not keep this. It's a lot of code
for one, and it also adds noticably to my load time. Not the end of
the world, because ideally I'm not starting Emacs *that* often, but
it's worth considering the implication.

The main thing slowing it down is the custom mode-line code, which is
hardly surprising.