Add gedcom.el to vendored files
This is a pretty old mode. Seems it used to be linked from the Emacs
Wiki but it's not there anymore. It's a file format for genealogy, I
thought it was kind of interesting and didn't want to lose track of
it.

Currently, I'm not requiring it in my config file because it's not
something that I'm going to use regularly.