Add d-mode to my Emacs config
I played with D a few nights ago doing some TDD code kata work. And I
was quietly impressed with it. It wasn't something that shook me to my
core, but it was a surprisingly pleasant little language.

It felt like something in between C++ and Java. Still able to do some
low level operations, but more expressive than C/C++ tend to be, hence
the Java comparison. But also without feeling as bloated as Java?

Kind of a long commit message for four lines, but I was actually
surprised by D. And I think I'll play with it again in the future.