Correct java configuration
use-package does allow for configuring built in package, however
"java-mode" lives in a file called "cc-mode". Which if the name is any
indication is a file named after C and C++.

I _think_ use-package allows for redundant calls, so I suppose I could
have a declaration up in the C family section, and another down with
Java. But for now, just use a plain old hook.

That after-save-hook is pretty janky. It shouldn't have to do that,
and it might not. I found myself running it a lot but I might have
just not gotten it?