Add a patch to fix issue with counsel-projectile mode
Recently there were some changes introduced in projectile that made it
so there weren't "default" keybindings. Along with that was apparently
a refactor that changed the names of some variables; and one that
counsel-projectile relied on is now gone. What I've put in here makes
it work, but hopefully this code will find it's way into the package
and I won't need to patch it like this.