Remove evil-mode from my Emacs config
This has been a long time coming. I've been feeling that Emacs and Vim
bindings don't mix well for a while. While evil is probably the best
Vim emulation out there, it conflicts with so much stuff it felt like
it was time to turn it off. I've left the use-package statement
commented out, because I'm unsure this is a decision I'm going to
stick with.