Update the prose in my Emacs config
A few of these either didn't really having anything written for them
or they had fallen out of date. I've also added a Preamble section
explaining why there is so much prose in this config in the first
place.

My hope is that if I talk to people who are interested in Emacs, that
they can look at this file and learn about how to go about setting up
a configuration; which is something I found to be harder than it
needed to be when I was trying to learn about it.