Remove the custom mode line code
Something changed in the spaceline package that this depended on, and
I don't feel like going and tracking it down.

There's a package that seems to do basically what I'd done, but I
wouldn't need to maintain it. Might check that out. For now I'm just
going to roll with the vanilla modeline.