Small updates
This is mostly a collection of spelling/grammar fixes. But I also put
in some default lists for smartparens so it will have better behavior
in some common languages.

Fixed the interactive prompt for counsel-projectile-rg, so now there's
actually a colon and space after it. Which is nicer to look at.

Finally, this adds a minor mode for highlighting indentation. Which is
nice for working with YAML