Ensure custom-file always exists
I've been putting off writing this fix for too long. Partially because
it's so simple I didn't even bother with it, but it's really high time
I dealt with it.

Just run a `touch` command on the custom-file before trying to load
it. That way if it's the initial bootstrapping of the system Emacs
won't barf trying to load a file that doesn't exist.