Overhaul Solargraph config in NeoVim
Ruby LSPs are tricky to manage. This has a lot to do with environment
management tools like rbenv. Automatic management solutions like Mason
tend to conflict with env tools, but the env tools are a layer of
indirection over executables.

Previously, I just assumed everything would be managed with rbenv and
hard-coded it. This conflicts with how management looks in NixOS where
rbenv is redundant.

So, I've overhauled the whole thing. Now there's a three step process to
find if Solargraph is available.
  1. Check `bundle`: this should cover basically every case
  2. Explicitly check `.rbenv/`
  3. Check if `solargraph` is in the $PATH