Switch to Xanmod kernel
I ran into a lot of issues when upgrading my system. It seems that the
current version of the Zen kernel has issues with Nvidia drivers right
now. Unsure what the current state of things are, but I read that
historically there have been issues with keeping consistent maintainers
of the Zen packages in Nix.

Zen *was* an improvement over the standard kernel, but it wasn't a
massive change. I've heard Xanmod is a significant improvement over the
standard Kernel and Zen; but it can have stability issues. I'm going to
try it out for now and see how it goes.

Initial benchmarking on the system *does* seem to be better so far. I'll
see how it lasts.