Archive NixOS config
I really enjoyed NixOS; and would love to give it another try some day.
However, I was running into far too many compatibility issues with very
bread-and-butter dev tooling for it to be worth my time. I think in the
future it could be worth another look. But not right now.